refactor: all-plugins.tsx => educational-elements.tsx
